+++ doc/CHANGELOG.md
@@ -29,6 +29,7 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
+* `#2743` Clear work packages filters when the work packages menu item is clicked
 * `#3072` Timelines rendering of top table border and text is slightly off
 * `#3108` [Work package tracking] Too many users selectable as watchers in public projects
 * `#3334` [CodeClimate] Mass Assignment WikiController